Switch back to the example alpine builder in CI
In CI a `pack build` is performed using one of libcnb's example CNBs
and the upstream `cnb/sample-builder:alpine` builder, in order to test
that our static MUSL compile works with non-libc base images.

However, in #594 the version of that builder had to be pinned to
an older release, to work around a temporary incompatibility
between the latest version of the builder and the latest stable
Pack CLI:
https://github.com/buildpacks/samples/issues/170

In addition, in #775 the image was switched to our builder image
since the example images didn't use a CNB lifecycle that was
compatible with Buildpack API 0.13.

Both issues have since been resolved, so we can switch back to
the latest version of the upstream example alpine builder,
giving us musl coverage again.

GUS-W-14086879.