Remove unnecessary usages of the `vec!` macro
For cases where a known-at-compile-time length array is being
passed to a function that accepts an `Into<Iterator>`, there is
no need to wrap the array in a `vec![]` invocation.

I spotted a couple of these in one of the newly added libcnb-test
tests, and a search and replace later it turns out there were quite
a few instances of this!