Replace state::Announce with ParagraphInspectWrite
We identified that there were gaps in the stateful implementation of `state::Announce`. It also leaked to the user and was in general not nice. To fix this, we've decided to inspect the values being printed to our writer using ParagraphInspectWrite.

When we write two newlines at the end of a write call then we know the last thing emitted was a "paragraph". We expose this information via `was_paragraph` associated function, and that allows us to simplify the code dramatically.

## Identified edge case in ParagraphInspectWrite

The first stub of ParagraphInspectWrite worked if an output emits two newlines directly like "Hello\n\n" but does not catch two writes like "Hello\n", "\n", which happens frequently as I use writeln! to add empty newlines to visually differentiate between only single lines written.

I added a test and fixed this issue by storing the number of newlines written since a non-newline character had been encountered.

## Original problem definition that this is solving

When we emit a line or paragraph we cannot know what the next output will be.
If two warnings are emitted in a row, or a warning and an error, etc. then
we do not want to emit two empty newlines. It sounds like a nit, but it looks
bad. We shouldn't do it.

Here are cases of the correct behavior:

- Warning after a step:
  - 🎉 Last output was a single empty newline, print a single newline
  - 🎉 Print the warning
  - 🎉 Print an empty newline
- Warning after another warning:
  - 🔇 Last output was a double newline, therefore we do not need to output any
    additional newlines.
  - 🎉 Print the warning
  - 🎉 Print an empty newline
- Warning after starting the buildpack:
  - 🔇 Header ends in an empty newline, don't print a leader newline
  - 🎉 Print the warning
  - 🎉 Print an empty newline