Fix header followed by warning tests
Header followed by a warning was returning an accidental extra newline.

This is caused due to the way that we're inspecting writes via `ParagraphInspectWrite`, it looks at the end of the string, counting newlines. However if there is a control (color) sequence or reset sequence it this will trigger it to also stop looking.

In the future we should make this more robust by accounting for any non-visible character, but in the short term I'm able to eliminate such results from being generated in the first place in the `colorize` function.