Adjust visibility due to clippy
We do not intend for anyone to directly use or implement the trait `AnnounceSupportedState`, it is now `pub(crate)`.

On the flip side, developers are expected to use items in the `state` module, so I made that `pub`.

Finally, clippy is giving us a warning: 

```
99  | / impl<S> BuildpackOutput<S>
100 | | where
101 | |     S: AnnounceSupportedState,
    | |______________________________^ implementation `buildpack_output::BuildpackOutput<S>` is reachable at visibility `pub`
    |
```

Which is what we want (I think). We want `write_mut` to be exercised indirectly through `warning`, `important`, and `error`. I ignored this warning with `allow(private_bounds)`.

I added a call to `warning()` in the doc test to ensure that it can still be called without the `AnnounceSupportedState` trait in bounds.