Refactor inject_default_ansi_escape to take enum
I was about to write docs about how this was an internal function was only known to work with escape codes we expect, then I realized that would mean my method signature was lying because it says it takes an &str. To make it not lie, I changed it to an enum which we control, and that let me delete the clarifying doc.