Add test for empty line case
Funny enough, the original logic to not append a space when the line is empty was added because my editor has "trim trailing whitespace" and it wasn't possible for me to make a heredoc test case with a trailing space (without changing my editor settings which I really didn't want to do).

In adding the test I'm now needing to revert https://github.com/heroku/libcnb.rs/commit/f620ce37b7c7dbcf245ecf113eccca6bcde68ed8#diff-811be8ca19faa4272e51de5beb3f1f9b088cfeeb65055d7a185dced3d071b5beL289 as we are asserting against that logic again.

It's good to have in there for future testing and consistency. It makes me wonder somewhat if that constraint (not trailing whitespace except for newlines) would be better enforced somewhere else, say in `prefix_lines` or friends. But I don't feel it's worth it to make that check everywhere if this is the only place it's needed. We can always do that later.