Document, test, and refactor prefix functions
There was a request that I document why the `contents.is_empty()` logic is present. I think it makes more logical sense to encode that information in the code, so I refactored to use a peekable iterator instead. Anyone who does not want to apply a prefix to an empty slice can do so via their custom function passed to `prefix_lines`.

I added docs for the `prefix_first_rest_lines` to call out the edge case of calling with an empty slice.