libcnb-test: Use `--trust-extra-buildpacks` with `pack build`
To force the builder to be trusted after changes were made in Pack
CLI v0.35.0 to improve security.

The new flag to restore the old behaviour was only added in Pack
v0.35.1, so using `libcnb-test` now requires that Pack version or
newer. (Adding a version check would require calling out to Pack
before every test execution, which is not worth it given that it only
affects tests, and this will be a one-off transition.)

Fixes #854.
GUS-W-16600037.