Update to Rust 2024 edition (#913)
Updates to the Rust 2024 edition, which is now available as of Rust 1.85:

- https://blog.rust-lang.org/2025/02/20/Rust-1.85.0.html
- https://doc.rust-lang.org/edition-guide/editions/transitioning-an-existing-project-to-a-new-edition.html
- https://doc.rust-lang.org/edition-guide/rust-2024/index.html

Most changes were applied via either `cargo clippy --fix` or `cargo fmt`
with the exception of a failure for the new `obfuscated_if_else`:
https://rust-lang.github.io/rust-clippy/master/index.html#obfuscated_if_else

I've also dropped the `module_name_repetitions` lint rule allow, since
that lint is no longer part of `clippy::pedantic` due to being too noisy,
so we no longer need to manually disable it.

GUS-W-17894359.