Use `--locked` with Cargo where possible
The Cargo `--locked` argument ensures that Cargo will fail with
an error if `Cargo.lock` is out of sync with `Cargo.toml`, rather
than the lockfile being silently updated.

Since this repo is a library, we currently don't commit its `Cargo.lock`
to Git (though we might want to revisit that for a few reasons in
the future).

However, there are still some places where we should be using
`--locked` that aren't already.

GUS-W-18062544.