Change default PATH delimiter
The default PATH delimiter is `` (empty string) rather than the OS delimiter. This has caused many issues and heartache https://github.com/buildpacks/spec/issues/285.

The upstream spec does not state the delimiter that should be used when none is provided. This change suggests that we should default to `:` for unix and `;` for windows. This can be seen as a dry-run for the spec proposal change https://github.com/buildpacks/rfcs/pull/326. If this change can be made with little to no reported problems, we can send that information upstream.

Other suggestions have been floated https://github.com/heroku/libcnb.rs/issues/534. I suggest that we prioritize the most common case, which for Heroku buildpacks is using the OS path separator.