Add a convenience function for retrying a download
When a download fails due to disk IO or if the file cannot be cleaned, do not suggest an (immediate) retry. This interface would allow code like this:

```retry
let mut download_attempts = 0;
while download_attempts <= MAX_RETRIES {
    match download_file(&artifact.url, &tarball_path) {
        Err(error) if error.retry_suggested() && download_attempts < MAX_RETRIES => {
            let sub_bullet = log_background_bullet.cancel(format!("{error}"));
            download_attempts += 1;
            thread::sleep(Duration::from_secs(1));
            log_background_bullet = sub_bullet.start_timer("Retrying");
        }
        result => {
            result?;
            let _ = log_background_bullet.done();
            break;
        }
    }
}
```

This allows developers who care less about the specific error states to defer the decision to whether or not an error can be retried to a canonical location.