Improved download client features
Introduces the following features to the `download_file` utility function:
- Configurable timeouts & retries
- OTel tracing of requests
- Sync/async usage
- Platform certificate store support
- Consistent output via `bullet_stream`

This replaces the usage of `ureq` as the HTTP client library so it wasn't possible to make this change fully backwards-compatible as the previous `DownloadError` enum leaked this internal implementation detail.

The `download_file` function is also marked as deprecated. It should be replaced with the following which offers a bit more flexibility:

```rust
get(uri.as_ref())
  .call_sync()
  .and_then(|res| res.download_to_file_sync(destination));
```