Fix doctest failures for libherokubuildpack::download::download_file
The response (and thus SHA-256 digest) for https://example.com/, used in the doctest, recently changed.

Instead, we can use a URL the contents of which we control, such as the LICENSE symlink file from inside this very repository.

Using raw.githubusercontent.com just in case the GitHub servers themselves are ever having hiccups.

GUS-W-20016948