Ensure tracing covers `buildpack.on_error()`
Lifts the trace guard to the `libcnb_runtime()` scope so that it's lifecycle is extended to the end of the function in order to cover `buildpack.on_error()`.

Due to the trace initialization needing buildpack information, this also required buildpack descriptor reading to be lifted into the same scope. This descriptor is then passed through to the `detect`/`build`. This leads to some duplication of error handling if the descriptor fails to be read.